Разбил скрипт на отдельные модули
